Changed some int into ushort, added cam_common_send_cmd(), removed some includes
git-svn-id: https://svn.streamboard.tv/oscam/branches/merlin@201 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ include/CAM/cryptoworks.h
@@ -1,7 +1,7 @@
 #ifndef __CAM_CRYPTOWORKS_H__
 #  define __CAM_CRYPTOWORKS_H__
 
-int cryptoworks_card_init(uchar *, int);
+int cryptoworks_card_init(uchar *, ushort);
 int cryptoworks_do_ecm(ECM_REQUEST *);
 int cryptoworks_do_emm(EMM_PACKET *);
 int cryptoworks_card_info();
