+++ include/CAM/viaccess.h
@@ -1,7 +1,7 @@
 #ifndef __CAM_VIACCESS_H__
 #  define __CAM_VIACCESS_H__
 
-int viaccess_card_init(uchar *, int);
+int viaccess_card_init(uchar *, ushort);
 int viaccess_do_ecm(ECM_REQUEST *);
 int viaccess_do_emm(EMM_PACKET *);
 int viaccess_card_info();
