+++ include/CAM/videoguard.h
@@ -1,7 +1,7 @@
 #ifndef __CAM_VIDEOGUARD_H__
 #  define __CAM_VIDEOGUARD_H__
 
-int videoguard_card_init(uchar *, int);
+int videoguard_card_init(uchar *, ushort);
 int videoguard_do_ecm(ECM_REQUEST *);
 int videoguard_do_emm(EMM_PACKET *);
 int videoguard_card_info();
